# frontend/src/lib/api/openapi.json
{
	"components": {
		"schemas": {
...
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": {
							"$ref": "#/components/schemas/ValidationError"
						},
						"title": "Detail",
						"type": "array"
					}
				},
				"title": "HTTPValidationError",
				"type": "object"
			},
...
			"ValidationError": {
				"properties": {
					"ctx": {
						"title": "Context",
						"type": "object"
					},
					"input": {
						"title": "Input"
					},
					"loc": {
						"items": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "integer"
								}
							]
						},
						"title": "Location",
						"type": "array"
					},
					"msg": {
						"title": "Message",
						"type": "string"
					},
					"type": {
						"title": "Error Type",
						"type": "string"
					}
				},
				"required": [
					"loc",
					"msg",
					"type"
				],
				"title": "ValidationError",
				"type": "object"
			},
...
		}
	},
	"info": {
		"description": "Self-hosted reader for collections of PDF periodicals.",
		"title": "Paperstand",
		"version": "0.1.0"
	},
	"openapi": "3.1.0",
	"paths": {
...
		"/api/issues/{issue_id}/file": {
			"get": {
				"description": "Stream an issue's PDF, honouring range and conditional requests.",
				"operationId": "issue_file",
				"parameters": [
					{
						"in": "path",
						"name": "issue_id",
						"required": true,
						"schema": {
							"title": "Issue Id",
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/pdf": {}
						},
						"description": "The whole document"
					},
					"206": {
						"content": {
							"application/pdf": {}
						},
						"description": "The requested range"
					},
					"304": {
						"description": "The client's copy is current"
					},
					"416": {
						"description": "The requested range is not satisfiable"
					},
					"422": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						},
						"description": "Validation Error"
					}
				},
				"summary": "Issue File",
				"tags": [
					"files"
				]
			},
			"head": {
				"description": "Stream an issue's PDF, honouring range and conditional requests.",
				"operationId": "issue_file_head",
				"parameters": [
					{
						"in": "path",
						"name": "issue_id",
						"required": true,
						"schema": {
							"title": "Issue Id",
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/pdf": {}
						},
						"description": "The whole document"
					},
					"206": {
						"content": {
							"application/pdf": {}
						},
						"description": "The requested range"
					},
					"304": {
						"description": "The client's copy is current"
					},
					"416": {
						"description": "The requested range is not satisfiable"
					},
					"422": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						},
						"description": "Validation Error"
					}
				},
				"summary": "Issue File",
				"tags": [
					"files"
				]
			}
		},
...
	}
}